:recycle: refactor(service) : factory class에게 객체 생성 로직 위임
