+++ src/main/java/ladder/factory/VerticalMapper.java
@@ -1,0 +1,12 @@
+package ladder.factory;
+
+import ladder.domain.HorizontalLine;
+import ladder.domain.VerticalLine;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public abstract class VerticalMapper {
+
+
+}
